+++ src/server/server.ts
@@ -44,6 +44,11 @@ function disposeInjector(injector: KrasInjector) {
   }
 }
 
+function isUnique<T>(value: T, index: number, self: Array<T>) {
+  // checks if the value is truthy and was not seen beforehand
+  return value && self.indexOf(value) === index;
+}
+
 export class MockServer extends MockServerCore implements KrasServer {
   readonly injectors: Array<KrasInjector> = [];
   readonly middlewares: Array<KrasMiddleware> = [];
@@ -54,7 +59,7 @@ export class MockServer extends MockServerCore implements KrasServer {
     super(config);
 
     this.logLevel = config.logLevel || 'error';
-    this.on('error', (e) => this.log('error', e));
+    this.on('error', e => this.log('error', e));
 
     if (config.api === false) {
       this.recorder.disable();
@@ -81,14 +86,14 @@ export class MockServer extends MockServerCore implements KrasServer {
   }
 }
 
-export function readKrasConfig(options?: ConfigurationOptions, file?: string) {
+export function readKrasConfig(options?: ConfigurationOptions, ...files: Array<string>) {
   const dir = options.dir ? resolve(currentDir, options.dir) : currentDir;
-  return mergeConfiguration(
-    options,
-    readConfiguration(homedir(), krasrc),
-    readConfiguration(dir, krasrc),
-    readConfiguration(dir, file !== krasrc && file),
-  );
+  const configurations = files
+    .filter(Boolean)
+    .map(file => resolve(dir, file))
+    .filter(isUnique)
+    .map(path => readConfiguration(path));
+  return mergeConfiguration(options, ...configurations);
 }
 
 export function buildKras(config?: Partial<KrasConfiguration>) {
@@ -122,7 +127,7 @@ export function runWithKras(cb: KrasRunner, config?: KrasRuntimeConfiguration) {
 }
 
 export function connectToCli(server: MockServer, canManage = true) {
-  server.on('open', (svc) => {
+  server.on('open', svc => {
     const port = chalk.green(svc.port);
     const protocol = svc.protocol;
     const server = `${protocol}://localhost:${port}`;
@@ -134,59 +139,59 @@ export function connectToCli(server: MockServer, canManage = true) {
     }
   });
 
-  server.on('close', (svc) => {
+  server.on('close', svc => {
     console.log(`Connection to server closed.`);
   });
 
-  server.on('user-connected', (msg) => {
+  server.on('user-connected', msg => {
     if (isDebug(server.logLevel)) {
       console.log(`${chalk.green('WS')} + ${chalk.white(info(msg.id))}`);
     }
   });
 
-  server.on('user-disconnected', (msg) => {
+  server.on('user-disconnected', msg => {
     if (isDebug(server.logLevel)) {
       console.log(`${chalk.green('WS')} - ${chalk.white(info(msg.id))}`);
     }
   });
 
-  server.on('message', (msg) => {
+  server.on('message', msg => {
     if (isDebug(server.logLevel)) {
       console.log(`${chalk.green('WS')} << ${chalk.white(info(msg.content))}`);
     }
   });
 
-  server.on('broadcast', (msg) => {
+  server.on('broadcast', msg => {
     if (isInfo(server.logLevel)) {
       console.log(`${chalk.green('WS')} >> ${chalk.white(info(msg.content))}`);
     }
   });
 
-  server.on('missing', (req) => {
+  server.on('missing', req => {
     if (isError(server.logLevel)) {
       console.log(`${chalk.yellow(req.method)} ${chalk.gray(req.target)}${chalk.white(req.url)}`);
     }
   });
 
-  server.on('request', (req) => {
+  server.on('request', req => {
     if (isDebug(server.logLevel)) {
       console.log(`${chalk.green(req.method)} ${chalk.gray(req.target)}${chalk.white(req.url)}`);
     }
   });
 
-  server.on('error', (msg) => {
+  server.on('error', msg => {
     if (isError(server.logLevel)) {
       console.log(`${chalk.red('ERR')} ${chalk.white(msg)}`);
     }
   });
 
-  server.on('debug', (msg) => {
+  server.on('debug', msg => {
     if (isDebug(server.logLevel)) {
       console.log(`${chalk.yellow('DBG')} ${chalk.white(msg)}`);
     }
   });
 
-  server.on('info', (msg) => {
+  server.on('info', msg => {
     if (isInfo(server.logLevel)) {
       console.log(`${chalk.bgWhite(chalk.black('INF'))} ${chalk.white(msg)}`);
     }
@@ -194,7 +199,7 @@ export function connectToCli(server: MockServer, canManage = true) {
 }
 
 export function runFromCli(options: ConfigurationOptions, rcfile: string) {
-  const config = readKrasConfig(options, rcfile);
+  const config = readKrasConfig(options, resolve(homedir(), krasrc), resolve(currentDir, krasrc), rcfile);
   const server = buildKrasWithCli(config);
   console.log(`Starting kras v${version} ...`);
   server.start();